Statically link to the Rust library
The Rust library was previously statically linked, which meant we had to install
the library to be able to run dependent application. Frow now, it will be
statically linked to the final binary, enhancing the developer’s and packager’s
expercience.

I’ve also added a Makefile to prebuild the library, and updated the instructions
in the README.md